Removed redundant cast
git-svn-id: https://svn.apache.org/repos/asf/incubator/pivot/trunk@754158 13f79535-47bb-0310-9956-ffa450edef68

+++ wtk/src/pivot/wtk/skin/terra/TerraPaletteSkin.java
@@ -181,7 +181,7 @@ public class TerraPaletteSkin extends WindowSkin {
 
         Font font = theme.getFont();
         titleLabel.getStyles().put("fontBold", true);
-        titleLabel.getStyles().put("fontSize", (int)Math.round(font.getSize2D() * 0.8f));
+        titleLabel.getStyles().put("fontSize", Math.round(font.getSize2D() * 0.8f));
         titleLabel.getStyles().put("color", titleBarColor);
 
         // Initialize the button flow pane
